dockerfile

+++ build.sh
@@ -3,6 +3,8 @@ set -e
 
 npm install
 
+mkdir -p cache
+
 # get sqlite3 version
 sqlite3_pkg=$(cat node_modules/sqlite3/package.json)
 sqlite3_version=$(node -e "console.log(JSON.parse(\`$sqlite3_pkg\`).version)")
@@ -10,10 +12,14 @@ sqlite3_version=$(node -e "console.log(JSON.parse(\`$sqlite3_pkg\`).version)")
 # download sqlite3 binary that need to pkg
 targets=("napi-v6-darwin-unknown-x64" "napi-v6-win32-unknown-x64" "napi-v6-linux-glibc-x64" "napi-v6-linux-glibc-arm64")
 for target in ${targets[@]}; do
-  wget https://github.com/TryGhost/node-sqlite3/releases/download/v$sqlite3_version/$target.tar.gz
-  tar -zxvf $target.tar.gz
+  if [ -f cache/$target.tar.gz ]; then
+    echo "$target.tar.gz has cache and skip to download."
+  else
+    wget -P cache https://github.com/TryGhost/node-sqlite3/releases/download/v$sqlite3_version/$target.tar.gz
+  fi
+
+  tar -zxvf cache/$target.tar.gz
   cp -rf $target ./node_modules/sqlite3/lib/binding
-  rm -rf $target.tar.gz
   rm -rf $target
 done
 
@@ -36,9 +42,13 @@ npm run pkg
 rm package.json
 mv package.json.bak package.json
 
-cd build
-tag=$(git describe --abbrev=0 --tags)
-for binary in ./*; do
-  tar zcvf $(echo $binary | sed -e 's/.exe//g')-$tag.tar.gz $binary
-  rm $binary
-done
+if test $1 = 'no-compress'; then
+  echo 'skip compression.'
+else
+  cd build
+  tag=$(git describe --abbrev=0 --tags)
+  for binary in ./*; do
+    tar zcvf $(echo $binary | sed -e 's/.exe//g')-$tag.tar.gz $binary
+    rm $binary
+  done
+fi
